* Something not working in reflections --> calcIndicesOfPrism has some bug --> not raysPerSample / raysPerDump

+++ include/reflection.h
@@ -2,6 +2,6 @@
 #define Reflection_H
 
 #include "mesh.h"
-__device__ int calcNextReflection(Point startPoint, Point endPoint, unsigned reflectionsLeft, int reflectionPlane, Point *reflectionPoint, float *reflectionAngle, Mesh *mesh);
+__device__ int calcNextReflection(Point startPoint, Point endPoint, unsigned reflectionsLeft, int reflectionPlane, Point *reflectionPoint, double *reflectionAngle, Mesh *mesh);
 
 #endif /* Reflection_H */
